# Simplify SpectralCoord implementation