--- a/astropy/coordinates/spectral_coordinate.py
+++ b/astropy/coordinates/spectral_coordinate.py
@@ -14,7 +14,17 @@ from astropy.utils.exceptions import AstropyUserWarning
 
 __all__ = ['SpectralCoord']
 
+
+class NoVelocityWarning(AstropyUserWarning):
+    pass
+
+
+class NiDistanceWarning(AstropyUserWarning):
+    pass
+
+
 KMS = u.km / u.s
+ZERO_VELOCITIES = CartesianDifferential(u.Quantity([0, 0, 0] * KMS))
 
 # Default distance to use for target when none is provided
 DEFAULT_DISTANCE = 1e6 * u.kpc
@@ -36,7 +46,7 @@ def _redshift_to_velocity(redshift):
     Convert a relativistic redshift to a velocity.
     """
     zponesq = (1 + redshift) ** 2
-    return c * (zponesq - 1) / (zponesq + 1)
+    return (c * (zponesq - 1) / (zponesq + 1)).to(u.km / u.s)
 
 
 def _relativistic_velocity_addition(vel1, vel2):
@@ -128,9 +138,16 @@ def attach_zero_velocities(coord):
     """
     Set the differentials to be stationary on a coordinate object.
     """
-    coord_diffs = CartesianDifferential(u.Quantity([0, 0, 0] * KMS))
+    coord_diffs = ZERO_VELOCITIES
     new_data = coord.cartesian.with_differentials(coord_diffs)
     return coord.realize_frame(new_data)
+
+
+def _get_velocities(coord):
+    if 's' in coord.data.differentials:
+        return coord.velocity
+    else:
+        return ZERO_VELOCITIES
 
 
 class SpectralCoord(SpectralQuantity):
@@ -219,96 +236,20 @@ class SpectralCoord(SpectralQuantity):
         # manually if either or both the observer and target are not
         # specified.
         if observer is None or target is None:
-            if radial_velocity is None and redshift is None:
+            if radial_velocity is None:
                 radial_velocity = getattr(value, 'radial_velocity', None)
 
-        # Keep track of whether the observer and target were specified
-        # explicitly or whether we use pseudo observers/targets (see below)
-        obj._observer_specified = observer is not None
-        obj._target_specified = target is not None
-
-        # Keep track of whether the radial velocity was explicitly specified
-        # (including as a redshift)
-        obj._rv_specified = radial_velocity is not None
-
-        if radial_velocity is None:
-            radial_velocity = 0 * u.km/u.s
-
-        if observer is not None:
-            observer = cls._validate_coordinate(observer, label='observer')
-
-        if target is not None:
-            target = cls._validate_coordinate(target, label='target')
-
-        # If no observer is defined, create a default observer centered in the
-        # ICRS frame. We never expose this to the user, and this is only used
-        # for internal purposes.
-        if observer is None:
-            if target is None:
-                observer = ICRS(ra=0 * u.degree, dec=0 * u.degree,
-                                pm_ra_cosdec=0 * u.mas/u.yr, pm_dec=0 * u.mas/u.yr,
-                                distance=1 * u.m, radial_velocity=0 * u.km/u.s)
-            else:
-                observer = SpectralCoord._create_coord_from_offset(target, radial_velocity)
-
-        # If no target is defined, create a default target with any provided
-        # redshift/radial velocities. We never expose this to the user, and
-        # this is only used for internal purposes.
-        if target is None:
-            target = SpectralCoord._create_coord_from_offset(observer, radial_velocity)
-
-        obj._observer = observer
-        obj._target = target
+        obj._radial_velocity = radial_velocity
+        obj._observer = cls._validate_coordinate(observer, label='observer')
+        obj._target = cls._validate_coordinate(target, label='target')
 
         return obj
 
     def __array_finalize__(self, obj):
         super().__array_finalize__(obj)
-        self._observer_specified = getattr(obj, '_observer_specified', None)
-        self._target_specified = getattr(obj, '_target_specified', None)
-        self._rv_specified = getattr(obj, '_rv_specified', None)
+        self._radial_velocity = getattr(obj, '_radial_velocity', None)
         self._observer = getattr(obj, '_observer', None)
         self._target = getattr(obj, '_target', None)
-
-    @staticmethod
-    @u.quantity_input(radial_velocity=u.km/u.s)
-    def _create_coord_from_offset(observer, radial_velocity):
-        """
-        Generates a default target or observer from a provided observer or
-        target with an offset defined such as to create the provided radial
-        velocity.
-
-        Parameters
-        ----------
-        observer : `~astropy.coordinates.BaseCoordinateFrame` or `~astropy.coordinates.SkyCoord`
-            Observer frame off which to base the target frame.
-        radial_velocity : `~astropy.units.Quantity`
-            Radial velocity used to calculate appropriate offsets between
-            provided observer and generated target.
-
-        Returns
-        -------
-        target : `~astropy.coordinates.BaseCoordinateFrame` or `~astropy.coordinates.SkyCoord`
-            Generated target frame.
-        """
-
-        # The generated observer or target will be set up along the same y and z
-        # coordinates as the target or observer, but offset along the x direction
-
-        observer_icrs = observer.transform_to(ICRS())
-
-        d = observer_icrs.cartesian.norm()
-        drep = CartesianRepresentation([DEFAULT_DISTANCE.to(d.unit),
-                                        0 * d.unit, 0 * d.unit])
-
-        obs_vel = observer_icrs.cartesian.differentials['s']
-
-        target = (observer_icrs.cartesian.without_differentials() + drep).with_differentials(
-            CartesianDifferential([_relativistic_velocity_addition(obs_vel.d_x, radial_velocity),
-                                   obs_vel.d_y.to(radial_velocity.unit),
-                                   obs_vel.d_z.to(radial_velocity.unit)]))
-
-        return observer_icrs.realize_frame(target)
 
     @staticmethod
     def _validate_coordinate(coord, label=''):
@@ -324,6 +265,10 @@ class SpectralCoord(SpectralQuantity):
         coord : `~astropy.coordinates.BaseCoordinateFrame`
             The new frame to be used for target or observer.
         """
+
+        if coord is None:
+            return
+
         if not issubclass(coord.__class__, BaseCoordinateFrame):
             if isinstance(coord, SkyCoord):
                 coord = coord.frame
@@ -334,13 +279,14 @@ class SpectralCoord(SpectralQuantity):
         # for generating differentials
         # TODO: change this to not set the distance and yield a warning once
         # there's a good way to address this in astropy.coordinates
-        if hasattr(coord, 'distance') and \
-                coord.distance.unit.physical_type == 'dimensionless':
+        with np.errstate(all='ignore'):
+            distance = getattr(coord, 'distance', None)
+        if distance is not None and distance.unit.physical_type == 'dimensionless':
             coord = SkyCoord(coord, distance=DEFAULT_DISTANCE)
             warnings.warn(
                 "Distance on coordinate object is dimensionless, an "
                 f"abritrary distance value of {DEFAULT_DISTANCE} will be set instead.",
-                AstropyUserWarning)
+                NiDistanceWarning)
 
         # If the observer frame does not contain information about the
         # velocity of the system, assume that the velocity is zero in the
@@ -349,7 +295,7 @@ class SpectralCoord(SpectralQuantity):
             warnings.warn(
                 "No velocity defined on frame, assuming {}.".format(
                     u.Quantity([0, 0, 0], unit=u.km/u.s)),
-                AstropyUserWarning)
+                NoVelocityWarning)
 
             coord = attach_zero_velocities(coord)
 
@@ -407,6 +353,12 @@ class SpectralCoord(SpectralQuantity):
             else:
                 value, unit = value.value, value.unit
 
+        if observer is not None:
+            observer = self._validate_coordinate(observer)
+
+        if target is not None:
+            target = self._validate_coordinate(target)
+
         value = value if value is not None else self.value
         unit = unit or self.unit
         observer = observer or self.observer
@@ -423,10 +375,12 @@ class SpectralCoord(SpectralQuantity):
         if (self.observer is None or self.target is None) and radial_velocity is None and redshift is None:
             radial_velocity = self.radial_velocity
 
-        return self.__class__(value=value, unit=unit,
-                              observer=observer, target=target,
-                              radial_velocity=radial_velocity, redshift=redshift,
-                              doppler_convention=doppler_convention, doppler_rest=doppler_rest, copy=False)
+        with warnings.catch_warnings():
+            warnings.simplefilter('ignore', NoVelocityWarning)
+            return self.__class__(value=value, unit=unit,
+                                  observer=observer, target=target,
+                                  radial_velocity=radial_velocity, redshift=redshift,
+                                  doppler_convention=doppler_convention, doppler_rest=doppler_rest, copy=False)
 
     @property
     def quantity(self):
@@ -452,8 +406,7 @@ class SpectralCoord(SpectralQuantity):
         `~astropy.coordinates.BaseCoordinateFrame`
             The astropy coordinate frame representing the observation.
         """
-        if self._observer_specified:
-            return self._observer
+        return self._observer
 
     @observer.setter
     def observer(self, value):
@@ -461,19 +414,11 @@ class SpectralCoord(SpectralQuantity):
         if self.observer is not None:
             raise ValueError("observer has already been set")
 
-        self._observer_specified = value is not None
+        self._observer = self._validate_coordinate(value, label='observer')
 
-        value = self._validate_coordinate(value, label='observer')
-
-        # The default target is based off the observer frame. In the case
-        # where both observer/target are initialized to defaults, and then
-        # the user sets a new observer, we need to create a new target based
-        # on the input frame to maintain rv/redshift continuity.
-        if self.target is None:
-            self._target = self._create_coord_from_offset(
-                value, self.radial_velocity)
-
-        self._observer = value
+        # Switch to auto-computing radial velocity
+        if self._target is not None:
+            self._radial_velocity = None
 
     @property
     def target(self):
@@ -485,19 +430,19 @@ class SpectralCoord(SpectralQuantity):
         `~astropy.coordinates.BaseCoordinateFrame`
             The astropy coordinate frame representing the target.
         """
-        if self._target_specified:
-            return self._target
+        return self._target
 
     @target.setter
     def target(self, value):
+
         if self.target is not None:
             raise ValueError("target has already been set")
 
-        self._target_specified = value is not None
+        self._target = self._validate_coordinate(value, label='target')
 
-        value = self._validate_coordinate(value, label='target')
-
-        self._target = value
+        # Switch to auto-computing radial velocity
+        if self._observer is not None:
+            self._radial_velocity = None
 
     @property
     def radial_velocity(self):
@@ -515,8 +460,14 @@ class SpectralCoord(SpectralQuantity):
         coordinate frame in that this calculates the radial velocity with
         respect to the *observer*, not the origin of the frame.
         """
-        return self._calculate_radial_velocity(self._observer, self._target,
-                                               as_scalar=True)
+        if self._observer is None or self._target is None:
+            if self._radial_velocity is None:
+                return 0 * KMS
+            else:
+                return self._radial_velocity
+        else:
+            return self._calculate_radial_velocity(self._observer, self._target,
+                                                   as_scalar=True)
 
     @property
     def redshift(self):
@@ -557,7 +508,7 @@ class SpectralCoord(SpectralQuantity):
         observer_icrs = observer.transform_to(ICRS())
         target_icrs = target.transform_to(ICRS())
 
-        pos_hat = SpectralCoord._norm_d_pos(observer_icrs, target_icrs)
+        pos_hat = SpectralCoord._normalized_position_vector(observer_icrs, target_icrs)
 
         d_vel = target_icrs.velocity - observer_icrs.velocity
 
@@ -569,7 +520,7 @@ class SpectralCoord(SpectralQuantity):
             return vel_mag * pos_hat
 
     @staticmethod
-    def _norm_d_pos(observer, target):
+    def _normalized_position_vector(observer, target):
         """
         Calculate the normalized position vector between two frames.
 
@@ -630,13 +581,13 @@ class SpectralCoord(SpectralQuantity):
         init_obs_vel = self._calculate_radial_velocity(self.observer, target)
         fin_obs_vel = self._calculate_radial_velocity(observer, target)
 
-        line_of_sight_unit_vec = self._norm_d_pos(observer, target)
+        line_of_sight_unit_vec = self._normalized_position_vector(observer, target)
 
         new_data = self._project_velocity_and_shift(init_obs_vel, fin_obs_vel, line_of_sight_unit_vec)
 
         new_coord = self.replicate(value=new_data,
-                                   observer=observer,
-                                   target=target)
+                                   observer=observer if observer is not self.observer else None,
+                                   target=target if target is not self.target else None)
 
         return new_coord
 
@@ -784,6 +735,7 @@ class SpectralCoord(SpectralQuantity):
             New spectral coordinate with the target/observer velocity changed
             to incorporate the shift.
         """
+
         if observer_shift is not None and (self.target is None or
                                            self.observer is None):
             raise ValueError("Both an observer and target must be defined "
@@ -799,11 +751,16 @@ class SpectralCoord(SpectralQuantity):
         #  assume it's a redshift float value and convert to velocity
 
         if target_shift is None:
+            if self._observer is None or self._target is None:
+                return self.replicate()
             target_shift = 0 * KMS
         else:
             target_shift = u.Quantity(target_shift)
             if target_shift.unit.physical_type == 'dimensionless':
                 target_shift = _redshift_to_velocity(target_shift)
+            if self._observer is None or self._target is None:
+                return self.replicate(value=_apply_relativistic_doppler_shift(self, target_shift),
+                                      radial_velocity=self._radial_velocity + target_shift)
 
         if observer_shift is None:
             observer_shift = 0 * KMS
@@ -815,10 +772,10 @@ class SpectralCoord(SpectralQuantity):
         target_icrs = self._target.transform_to(ICRS())
         observer_icrs = self._observer.transform_to(ICRS())
 
-        pos_hat = SpectralCoord._norm_d_pos(observer_icrs, target_icrs)
+        pos_hat = SpectralCoord._normalized_position_vector(observer_icrs, target_icrs)
 
-        target_velocity = target_icrs.velocity + target_shift * pos_hat
-        observer_velocity = observer_icrs.velocity + observer_shift * pos_hat
+        target_velocity = _get_velocities(target_icrs) + target_shift * pos_hat
+        observer_velocity = _get_velocities(observer_icrs) + observer_shift * pos_hat
 
         target_velocity = CartesianDifferential(target_velocity.xyz)
         observer_velocity = CartesianDifferential(observer_velocity.xyz)
@@ -834,7 +791,7 @@ class SpectralCoord(SpectralQuantity):
         init_obs_vel = self._calculate_radial_velocity(observer_icrs, target_icrs)
         fin_obs_vel = self._calculate_radial_velocity(new_observer, new_target)
 
-        line_of_sight_unit_vec = self._norm_d_pos(observer_icrs, target_icrs)
+        line_of_sight_unit_vec = self._normalized_position_vector(observer_icrs, target_icrs)
 
         new_data = self._project_velocity_and_shift(init_obs_vel, fin_obs_vel, line_of_sight_unit_vec)
 
@@ -885,7 +842,7 @@ class SpectralCoord(SpectralQuantity):
             target_repr = indent(repr(self.target), 12 * ' ').lstrip()
             repr_items.append(f'    target: {target_repr}')
 
-        if (self._observer_specified and self._target_specified) or self._rv_specified:
+        if (self._observer is not None and self._target is not None) or self._radial_velocity is not None:
             if self.observer is not None and self.target is not None:
                 repr_items.append('    observer to target (computed from above):')
             else: