… CHANGES.rst
… docs/coordinates/spectralcoord.rst