The current SpectralCoord implementation relies on an internal 'fake' observer and target when these are not specified explicitly, but I believe this actually makes the implementation more complicated than it needs to be (the original intent was to actually make it simpler but this is not the case in the end!).

In addition, some of the implementation is also overly complex because of an original use case (changing the actual observer location and taking into account parallax effects) which didn't make it into the final API since we determined that this might not be needed. This PR shows how we can simplify the implementation while still having the tests pass (at least they do for me locally).

Even if we do eventually implement more complex use cases, I think we should keep the implementation as simple as possible for the current use cases, for maintainability and speed.

The resulting code in this PR has fewer frame transformations and vector operations, and is therefore faster. For instance:

**Before**

```python
In [1]: from astropy.coordinates import SpectralCoord                                                                                                                            

In [2]: from astropy import units as u                                                                                                                                           

In [3]: velocity = 100 * u.km / u.s                                                                                                                                              

In [4]: %timeit sc = SpectralCoord(100, 'GHz')                                                                                                                                   
2.77 ms ± 121 µs per loop (mean ± std. dev. of 7 runs, 100 loops each)

In [5]: sc = SpectralCoord(100, 'GHz')                                                                                                                                   

In [6]: %timeit res = sc.with_radial_velocity_shift(velocity)                                                                                                                    
7.11 ms ± 410 µs per loop (mean ± std. dev. of 7 runs, 100 loops each)
```

**After**

```python
In [1]: from astropy.coordinates import SpectralCoord                                                                                                                            
^[[A^[[A
In [2]: from astropy import units as u                                                                                                                                           

In [3]: velocity = 100 * u.km / u.s                                                                                                                                              

In [4]: %timeit sc = SpectralCoord(100, 'GHz')                                                                                                                                   
134 µs ± 2.37 µs per loop (mean ± std. dev. of 7 runs, 10000 loops each)

In [5]: sc = SpectralCoord(100, 'GHz')                                                                                                                                           

In [6]: %timeit res = sc.with_radial_velocity_shift(velocity)                                                                                                                    
418 µs ± 34.4 µs per loop (mean ± std. dev. of 7 runs, 1000 loops each)
```

So the initialization in this case is 20x faster and the call to ``with_radial_velocity_shift`` is 17x faster. This is just an example of course, but overall I think all operations will be the same or faster.

This will require a careful review to make sure I haven't actually broken anything. It might be possible to simplify things further but I haven't had a chance to work on this more as I'm not going to be working much on astropy for a couple of months. So if anyone would like to review this in detail and push commits to it, please feel free to do so!

(cc @mhvk who originally commented on the SpectralCoord PR that the internal coordinates seemed unnecessary)