@astrofrog 2020-05-28
@mhvk - thanks for the detailed review! I think I've addressed all your comments

@nmearl - thanks for the suggestions! I've included them here.