@astrofrog 2020-10-16
I've rebased this - I'd really like to get this in v4.2 if possible.

@mhvk @nmearl - are you happy with this now?

Anyone else want to review this? (@eteq?)