@mhvk 2020-10-16
@astrofrog - I think there is just one outstanding comment... Otherwise looks good!