@astrofrog 2020-10-17
Ok thanks! I'll look into fixing the CI today.