@Cadair 2020-10-23
@astrofrog I rebased to fix the merge conflict and fixed the flake8 issue :smile: