@astrofrog 2020-10-23
@StuartLittlefair @adrn @astrojuanlu @eteq @mhvk - it would be best for this to be merged by a coordinates maintainer, so I'll leave it up to you to merge. I think it'd be a shame to miss v4.2 though so please merge today if possible! 🙏