@StuartLittlefair 2020-10-23
Hi @astrofrog - since @mhvk has run his fine tooth comb over this PR I'm happy to merge in principle, but I notice from the coverage that the cases where, for example, ```self._target``` and ```self._observer``` are both None are not being tested. There are a number of these lines. 

Maybe worth looking to see if the code is well tested in these instances before hitting the button?