@astrofrog 2020-10-23
@StuartLittlefair - I've tried to improve the test coverage, now 97% for the ``spectral_coordinate.py`` file. I won't have time to get to 100% today but hopefully this should be good enough?