@pllim 2020-10-23
Tsk tsk... Should have checked the remote data job. Can you please fix the failing doctest?

```
_________________________ [doctest] spectralcoord.rst __________________________
310 method. This method can take the name of an existing coordinate/velocity frame,
311 a :class:`~astropy.coordinates.BaseCoordinateFrame` instance, or any arbitrary
312 3D position and velocity coordinate object defined either as a
313 :class:`~astropy.coordinates.BaseCoordinateFrame` or a |SkyCoord| object. Most
314 commonly-used frames are accessible using strings. For example to transform to a
315 velocity frame stationary with respect to the center of the Earth (so removing
316 the effect of the Earth's rotation), we can use the ``'gcrs'`` which stands for
317 *Geocentric Celestial Reference System* (GCRS)::
318 
319     >>> sc_ttau.with_observer_stationary_relative_to('gcrs')  # doctest: +FLOAT_CMP +REMOTE_DATA
Differences (unified diff with -expected +actual):
    @@ -1,7 +1,7 @@
    -<SpectralCoord
    +<SpectralCoord
        (observer: <GCRS Coordinate (obstime=2019-04-24T02:32:10.000, obsgeoloc=(0., 0., 0.) m, obsgeovel=(0., 0., 0.) m / s): (x, y, z) in m
    -                  (-5878853.86171412, -192921.84773269, -2470794.19765021)
    +                  (-5878853.86172938, -192921.84773269, -2470794.19765021)
                    (v_x, v_y, v_z) in km / s
    -                  (4.33251262e-09, 8.96175625e-08, -1.49258412e-08)>
    +                  (-1.09244138e-08, 8.96161946e-08, -1.49267726e-08)>
         target: <ICRS Coordinate: (ra, dec, distance) in (deg, deg, pc)
                     (65.497625, 19.53511111, 144.321)
    @@ -9,6 +9,6 @@
                     (1.37949782e-15, 1.46375638e-15, 23.9)>
         observer to target (computed from above):
    -      radial_velocity=40.674086368345165 km / s
    -      redshift=0.00013568335316072044)
    +      radial_velocity=40.67408630725199 km / s
    +      redshift=0.0001356833529568835)
       [200.00024141, 210.00025348, 220.00026555, 230.00027762, 240.00028969,
        250.00030176, 260.00031383, 270.0003259 , 280.00033797, 290.00035004,
/home/travis/build/astropy/astropy/docs/coordinates/spectralcoord.rst:319: DocTestFailure
```

https://travis-ci.com/github/astropy/astropy/jobs/404743171 (This PR)

https://travis-ci.com/github/astropy/astropy/jobs/404850035 (On `master` after merge)