@astrofrog 2020-10-23
Huh wait but I did check it locally and that's what led to the doctest updates in https://github.com/astropy/astropy/pull/10398/commits/26442d19cdfc87f3492f21cc638694145f525b03 - I wonder why the results are different on Travis?!