@pllim 2020-10-23
> I wonder why the results are different on Travis?!

The same job also uses numpy dev, though not sure if that is the cause or not. If only we can use https://docs.travis-ci.com/user/running-build-in-debug-mode/