@mhvk 2020-10-23
Looks like some rounding error problem - these velocities are equal to 0 km/s within a few μm/s... Just skip the test until we have a `FLOAT_CMP` where we also can give absolute tolerances?